@@ -7,7 +7,7 @@ const messages = {
   experiences: 'Expériences',
   firstName: 'Prénom',
   footerText:
-    'Ce site est garanti 100% sans cookies ! Développé par Sébastien CAYET (lui-même) avec Angular 16.',
+    'Ce site est garanti 100% sans cookies ! Développé par Sébastien CAYET (lui-même) avec Angular 18.',
   from: 'De',
   greatings: 'Bonjour !',
   greatingsDescription:
@@ -46,6 +46,7 @@ const messages = {
   theory: 'Théorie',
   to: 'à',
   tooling: 'Outils',
+  visit: 'Visiter',
 };
 
 export default messages;
