@@ -7,7 +7,7 @@ const messages = {
   experiences: '経験',
   firstName: 'なまえ',
   footerText:
-    'このサイトは100％クッキーフリーを保証しています! CAYET Sébastien （自身）がAngular 16で開発。',
+    'このサイトは100％クッキーフリーを保証しています! CAYET Sébastien （自身）がAngular 18で開発。',
   from: '',
   greatings: 'こんにちは!',
   greatingsDescription: '私はセバスチャン. ウェブデベロッパーになって4年。',
@@ -46,6 +46,7 @@ const messages = {
   to: 'から',
   toAfter: 'まで',
   tooling: 'ツール',
+  visit: '見る',
 };
 
 export default messages;
